Boss1
added health to enemybase. Added boss 1 (Summon with L). Acts like standard enemy, but can start a duel with M once it's health drops to a quarter. (Increment health with Y for testing). When it is at a quarter or less health you can start a duel that if succesful will kill the boss, and if unsuccesful will heal the boss partially. Console shows boss health when incrementing and after losing a duel.